python/cluster/requirements.txt
cassandra-driver==3.2.1
joblib
pathlib2
psycopg2==2.6.2